+++ ccd-adapter/src/test/java/uk/gov/hmcts/cmc/ccd/mapper/CaseMapperTest.java
@@ -84,6 +84,7 @@ public class CaseMapperTest {
 
         //then
         assertThat(claim).isEqualTo(ccdCase);
+        assertEquals(MapperUtil.getMediationOutcome(ccdCase), claim.getMediationOutcome().orElse(null));
     }
 
     @Test
