+++ ccd-sample-data/src/main/java/uk/gov/hmcts/cmc/ccd/sample/data/SampleData.java
@@ -82,8 +82,8 @@ import static uk.gov.hmcts.cmc.domain.models.particulars.DamagesExpectation.THOU
 
 public class SampleData {
 
-    public static final String AMOUNT = "12398";
-    public static final String SUBMIT_MORE_DOCS_INSTRUCTION = "submit more docs";
+    private static final String AMOUNT = "12398";
+    private static final String SUBMIT_MORE_DOCS_INSTRUCTION = "submit more docs";
     public static final String MANCHESTER_CIVIL_JUSTICE_CENTRE_CIVIL_AND_FAMILY_COURTS
         = "Manchester Civil Justice Centre (Civil and Family Courts)";
 
@@ -737,7 +737,7 @@ public class SampleData {
             .build();
     }
 
-    public static CCDCase getCCDCitizenCaseWithRespondent(CCDRespondent respondent){
+    public static CCDCase getCCDCitizenCaseWithRespondent(CCDRespondent respondent) {
         List<CCDCollectionElement<CCDApplicant>> applicants
             = singletonList(CCDCollectionElement.<CCDApplicant>builder().value(getCCDApplicantIndividual()).build());
         List<CCDCollectionElement<CCDRespondent>> respondents
