+++ domain-model/src/main/java/uk/gov/hmcts/cmc/domain/models/Claim.java
@@ -247,7 +247,7 @@ public class Claim {
         return Optional.ofNullable(channel);
     }
 
-    public Optional<MediationOutcome> getMediationOutcome(){
+    public Optional<MediationOutcome> getMediationOutcome() {
         return Optional.ofNullable(mediationOutcome);
     }
