+++ src/main/java/uk/gov/hmcts/cmc/claimstore/utils/ResponseHelper.java
@@ -6,6 +6,8 @@ import uk.gov.hmcts.cmc.domain.models.response.ResponseType;
 
 import java.util.function.Predicate;
 
+import static uk.gov.hmcts.cmc.domain.models.response.ResponseType.FULL_DEFENCE;
+import static uk.gov.hmcts.cmc.domain.models.response.ResponseType.PART_ADMISSION;
 import static uk.gov.hmcts.cmc.domain.models.response.YesNoOption.YES;
 
 public class ResponseHelper {
@@ -13,6 +15,10 @@ public class ResponseHelper {
     private ResponseHelper() {
         // Utility class, no instances
     }
+
+    public static final Predicate<Response> isResponsePartOrFullDefence = response ->
+        response.getResponseType() == PART_ADMISSION
+            || response.getResponseType() == FULL_DEFENCE;
 
     public static boolean admissionResponse(Response response) {
         if (response != null && (response.getResponseType().equals(ResponseType.PART_ADMISSION)
@@ -29,7 +35,8 @@ public class ResponseHelper {
 
     public static boolean isOptedForMediation(Response response) {
         return response.getFreeMediation()
-            .filter(Predicate.isEqual(YES))
+            .filter(YES::equals)
             .isPresent();
     }
+
 }
